+++ app/src/main/kotlin/com/blocksdecoded/coinwave/data/crypto/network/CryptoNetworkClient.kt
@@ -8,8 +8,11 @@ import retrofit2.http.Query
 // Created by askar on 7/19/18.
 interface CryptoNetworkClient {
     @GET(CryptoConfig.CURRENCIES_PATH)
-    fun getCurrencies (
+    fun getCurrencies(@Query("limit") limit: Int): Call<CurrencyListResponse>
+
+    @GET(CryptoConfig.CURRENCIES_PATH)
+    fun getCurrencies(
             @Query("limit") limit: Int,
-            @Query("ids") ids: String = ""
+            @Query("ids") ids: String
     ): Call<CurrencyListResponse>
 }